% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{disconnect_taxa_db}
\alias{disconnect_taxa_db}
\title{Disconnect a taxadb connection}
\usage{
disconnect_taxa_db(db)
}
\arguments{
\item{db}{A connection returned by \code{\link[=get_taxa_db]{get_taxa_db()}}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Disconnect a taxadb connection
}
\keyword{internal}
