% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{empty_candidates}
\alias{empty_candidates}
\title{Create an empty candidate table}
\usage{
empty_candidates()
}
\value{
A zero-row candidate data frame.
}
\description{
Create an empty candidate table
}
\keyword{internal}
