% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{empty_raw_matches}
\alias{empty_raw_matches}
\title{Create an empty raw match table}
\usage{
empty_raw_matches()
}
\value{
A zero-row data frame with taxonomy columns.
}
\description{
Create an empty raw match table
}
\keyword{internal}
