% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{empty_taxonomy}
\alias{empty_taxonomy}
\title{Create an empty taxonomy table}
\usage{
empty_taxonomy()
}
\value{
A zero-row data frame with the ITIS fields used by capemlTaxa.
}
\description{
Create an empty taxonomy table
}
\keyword{internal}
