% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{get_itis_data}
\alias{get_itis_data}
\title{Retrieve ITIS data from the local database}
\usage{
get_itis_data()
}
\value{
A data frame containing the ITIS taxonomic data.
}
\description{
Helper function to retrieve the ITIS table as a data frame from the local
taxadb database.
}
\keyword{internal}
