% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{get_taxa_db}
\alias{get_taxa_db}
\title{Connect to the local taxadb database}
\usage{
get_taxa_db()
}
\value{
A DBI-compatible taxadb connection.
}
\description{
Connect to the local taxadb database
}
\keyword{internal}
