% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{get_taxadb_tbl_name}
\alias{get_taxadb_tbl_name}
\title{Get the name of the ITIS table in the database}
\usage{
get_taxadb_tbl_name(connection, provider = "itis")
}
\arguments{
\item{connection}{A DBI database connection.}

\item{provider}{The provider name (default "itis").}
}
\value{
The name of the ITIS table as a character string.
}
\description{
Helper function to determine the correct ITIS table name in the database.
}
\keyword{internal}
