% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{normalize_taxon_names}
\alias{normalize_taxon_names}
\title{Normalize names for ITIS lookup}
\usage{
normalize_taxon_names(names, match_by)
}
\arguments{
\item{names}{Character vector of organism names.}

\item{match_by}{Either \code{"scientific"} or \code{"common"}.}
}
\value{
A normalized character vector of the same length as \code{names}.
}
\description{
Normalize names for ITIS lookup
}
\keyword{internal}
