% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{provision_taxa_db}
\alias{provision_taxa_db}
\title{Provision the ITIS database views}
\usage{
provision_taxa_db(db)
}
\arguments{
\item{db}{An open taxadb connection.}
}
\value{
\code{db}, invisibly, after creating the ITIS Darwin Core and common-name
views.
}
\description{
Provision the ITIS database views
}
\keyword{internal}
