% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{query_accepted_taxa}
\alias{query_accepted_taxa}
\title{Query accepted ITIS taxa}
\usage{
query_accepted_taxa(ids, db)
}
\arguments{
\item{ids}{Accepted ITIS identifiers.}

\item{db}{An open taxadb connection.}
}
\value{
A data frame containing canonical accepted records.
}
\description{
Query accepted ITIS taxa
}
\keyword{internal}
