% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{query_itis_names}
\alias{query_itis_names}
\title{Query ITIS names}
\usage{
query_itis_names(names, match_by, db)
}
\arguments{
\item{names}{Normalized names to query.}

\item{match_by}{Either \code{"scientific"} or \code{"common"}.}

\item{db}{An open taxadb connection.}
}
\value{
A data frame of raw ITIS matches.
}
\description{
Query ITIS names
}
\keyword{internal}
