% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{resolve_candidates}
\alias{resolve_candidates}
\title{Resolve ambiguous candidate taxa}
\usage{
resolve_candidates(candidates, path)
}
\arguments{
\item{candidates}{Canonical candidate table from \code{\link[=build_candidates]{build_candidates()}}.}

\item{path}{Directory containing resolution files.}
}
\value{
A list containing marked candidates and resolution-status vectors.
}
\description{
Resolve ambiguous candidate taxa
}
\keyword{internal}
