% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{validate_write_taxa_map_inputs}
\alias{validate_write_taxa_map_inputs}
\title{Validate taxa map inputs}
\usage{
validate_write_taxa_map_inputs(taxa_df, path)
}
\arguments{
\item{taxa_df}{A prospective input data frame.}

\item{path}{Prospective output directory.}
}
\value{
\code{NULL}, invisibly. Throws an error for invalid inputs.
}
\description{
Validate taxa map inputs
}
\keyword{internal}
