% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{write_duplicate_candidates}
\alias{write_duplicate_candidates}
\title{Write ambiguous candidate taxa}
\usage{
write_duplicate_candidates(candidates, path)
}
\arguments{
\item{candidates}{Marked ambiguous candidate rows.}

\item{path}{Output directory.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Write ambiguous candidate taxa
}
\keyword{internal}
